DAO added.

+++ src/main/java/org/training/issuetracker/db/dao/interfaces/IUserDAO.java
@@ -6,7 +6,7 @@ import org.training.issuetracker.db.beans.User;
 
 public interface IUserDAO {
 	public List<User> getAll() throws Exception;
-	public User getById(int id) throws Exception;
+	public User getById(long id) throws Exception;
 	public void add(User user) throws Exception;
 	public User getByEmailAndPassword(String email, String password) throws Exception;
 }
